Enable 64 bit node IDs.

+++ src/types.rs
@@ -6,7 +6,7 @@ use std;
 
 use malloc_size_of::MallocSizeOf;
 
-pub type NodeID = u32;
+pub type NodeID = u64;
 
 #[derive(Serialize, Deserialize, Default, Eq, PartialEq, PartialOrd, Ord, Clone, Debug, MallocSizeOf, Hash)]
 #[repr(C)]
